+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Allow to filter observations per basis of record
 - By default, (at page load), only unseen observations are shown
+- Allow to filter observations per validation status
 - Updated backend and frontend dependencies
 - Internal improvements (https://github.com/riparias/gbif-alert/issues/282)