+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Allow to filter observations per basis of record
 - By default, (at page load), only unseen observations are shown
 - Allow to filter observations per validation status
+- Multiple user interface improvements (https://github.com/riparias/gbif-alert/issues/145, https://github.com/riparias/gbif-alert/issues/296, ...)
 - Updated backend and frontend dependencies
 - Internal improvements (https://github.com/riparias/gbif-alert/issues/282)
 
@@ -19,7 +20,7 @@
 
 # v1.7.7 (2024-11-07)
 
-- Fixed a bg with the maps API (see https://github.com/riparias/gbif-alert/issues/283)
+- Fixed a bug with the maps API (see https://github.com/riparias/gbif-alert/issues/283)
 
 # v1.7.6 (2024-07-26)